--- a/Cargo.toml
+++ b/Cargo.toml
@@ -44,8 +44,6 @@ merlin = "3.0.0"
 curve25519-dalek = {version = "3.2.0", features = ["serde"]}
 serde = { version = "1.0.124", features = ["derive"] }
 bincode = "1.2.1"
-tracing = "0.1"
-tracing-subscriber = "0.3"
 smallvec = "1.9.0"
 libc = "0.2"
 ark-curve25519 = { path = "./third_party/curve25519" }