--- a/circ_fields/src/lib.rs
+++ b/circ_fields/src/lib.rs
@@ -11,18 +11,29 @@ pub mod moduli {
     pub use super::ff_field::{F_BLS12381_FMOD_ARC, F_BN254_FMOD_ARC};
 }
 
+use ark_curve25519::Fr;
+use ark_ff::PrimeField;
+use ark_ff::{BigInt, BigInteger, Field as ark_Field};
 use ff_field::{FBls12381, FBn254};
 use ff_field::{F_BLS12381_FMOD, F_BN254_FMOD};
 use ff_field::{F_BLS12381_FMOD_ARC, F_BN254_FMOD_ARC};
 use int_field::IntField;
+use lazy_static::lazy_static;
+use num_traits::Zero;
 
 use ff::Field;
 use paste::paste;
+use rand::distributions::Standard;
+use rand::prelude::Distribution;
+use rug::integer::Order;
 use rug::Integer;
+use serde::de::Visitor;
+use serde::ser::Error;
 use serde::{Deserialize, Serialize};
 use std::fmt::{self, Display, Formatter};
 use std::ops::{Add, AddAssign, Mul, MulAssign, Neg, Sub, SubAssign};
 use std::sync::Arc;
+use ark_serialize::{CanonicalSerialize, CanonicalDeserialize};
 
 // TODO: rework this using macros?
 
@@ -33,13 +44,23 @@ pub enum FieldT {
     FBls12381,
     /// BN-254 scalar field as `ff`
     FBn254,
+    /// FCurve25519
+    FCurve25519,
     /// Generic field element based on `rug::Integer`
     IntField(Arc<Integer>),
+}
+
+lazy_static! {
+    /// Field modulus
+    pub static ref F_CURVE25519_FMOD: Integer = bigint_to_int(Fr::MODULUS);
+    /// Field modulus arc
+    pub static ref F_CURVE25519_FMOD_ARC: Arc<Integer> = Arc::new(F_CURVE25519_FMOD.clone());
 }
 
 impl Display for FieldT {
     fn fmt(&self, f: &mut Formatter) -> fmt::Result {
         match self {
+            Self::FCurve25519 => write!(f, "FieldT::FCurve25519"),
             Self::FBls12381 => write!(f, "FieldT::FBls12381"),
             Self::FBn254 => write!(f, "FieldT::FBn254"),
             Self::IntField(m) => write!(f, "FieldT::(mod {})", &*m),
@@ -71,6 +92,7 @@ impl FieldT {
         match m {
             m if m == &*F_BLS12381_FMOD => Some(Self::FBls12381),
             m if m == &*F_BN254_FMOD => Some(Self::FBn254),
+            m if m == &*F_CURVE25519_FMOD => Some(Self::FCurve25519),
             _ => None,
         }
     }
@@ -79,6 +101,7 @@ impl FieldT {
     #[inline]
     pub fn modulus(&self) -> &Integer {
         match self {
+            Self::FCurve25519 => &*F_CURVE25519_FMOD,
             Self::FBls12381 => &*F_BLS12381_FMOD,
             Self::FBn254 => &*F_BN254_FMOD,
             Self::IntField(m) => m.as_ref(),
@@ -89,6 +112,7 @@ impl FieldT {
     #[inline]
     pub fn modulus_arc(&self) -> Arc<Integer> {
         match self {
+            Self::FCurve25519 => F_CURVE25519_FMOD_ARC.clone(),
             Self::FBls12381 => F_BLS12381_FMOD_ARC.clone(),
             Self::FBn254 => F_BN254_FMOD_ARC.clone(),
             Self::IntField(m) => m.clone(),
@@ -99,6 +123,7 @@ impl FieldT {
     #[inline]
     pub fn default_value(&self) -> FieldV {
         match self {
+            Self::FCurve25519 => FieldV::FCurve25519(FCurve25519(Fr::zero())),
             Self::FBls12381 => FieldV::FBls12381(Default::default()),
             Self::FBn254 => FieldV::FBn254(Default::default()),
             Self::IntField(m) => FieldV::IntField(IntField::new(Integer::from(0), m.clone())),
@@ -134,8 +159,53 @@ pub enum FieldV {
     FBls12381(FBls12381),
     /// BN-254 scalar field element as `ff`
     FBn254(FBn254),
+    /// FCurve25519
+    FCurve25519(FCurve25519),
     /// Generic field element based on `rug::Integer`
     IntField(IntField),
+}
+
+#[derive(PartialEq, Eq, Clone, Debug, PartialOrd, Ord, Hash)]
+/// field for curve25519
+pub struct FCurve25519(pub Fr);
+
+impl Serialize for FCurve25519 {
+    fn serialize<S>(&self, serializer: S) -> Result<S::Ok, S::Error>
+    where
+        S: serde::Serializer,
+    {
+        let mut bytes = Vec::<u8>::new();
+        if let Err(_) = self.0.serialize_uncompressed(&mut bytes) {
+            return Err(S::Error::custom("serialize uncompressed error!"))
+        };
+        serializer.serialize_bytes(&bytes)
+    }
+}
+
+impl<'de> Deserialize<'de> for FCurve25519 {
+    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
+    where
+        D: serde::Deserializer<'de>,
+    {
+        struct MyFrVisitor;
+
+        impl<'de> Visitor<'de> for MyFrVisitor {
+            type Value = FCurve25519;
+            fn visit_bytes<E>(self, v: &[u8]) -> Result<Self::Value, E>
+            where
+                E: serde::de::Error,
+            {
+                let f: Fr = CanonicalDeserialize::deserialize_uncompressed(v).unwrap();
+                Ok(FCurve25519(f))
+            }
+
+            fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
+                formatter.write_str("struct MyFr")
+            }
+        }
+
+        deserializer.deserialize_bytes(MyFrVisitor)
+    }
 }
 
 impl FieldV {
@@ -143,6 +213,7 @@ impl FieldV {
     #[inline]
     pub fn ty(&self) -> FieldT {
         match self {
+            Self::FCurve25519(_) => FieldT::FCurve25519,
             Self::FBls12381(_) => FieldT::FBls12381,
             Self::FBn254(_) => FieldT::FBn254,
             Self::IntField(i) => FieldT::IntField(i.modulus_arc()),
@@ -175,6 +246,7 @@ impl FieldV {
     #[inline]
     pub fn modulus(&self) -> &Integer {
         match self {
+            Self::FCurve25519(_) => &*F_CURVE25519_FMOD,
             Self::FBls12381(_) => &*F_BLS12381_FMOD,
             Self::FBn254(_) => &*F_BN254_FMOD,
             Self::IntField(i) => i.modulus(),
@@ -186,6 +258,7 @@ impl FieldV {
     #[inline]
     pub fn recip_ref(&self) -> Self {
         match self {
+            Self::FCurve25519(pf) => Self::FCurve25519(FCurve25519(pf.0.inverse().unwrap())),
             Self::FBls12381(pf) => Self::FBls12381(pf.invert().unwrap()),
             Self::FBn254(pf) => Self::FBn254(pf.invert().unwrap()),
             Self::IntField(i) => Self::IntField(i.clone().recip()),
@@ -197,6 +270,7 @@ impl FieldV {
     #[inline]
     pub fn recip(self) -> Self {
         match self {
+            Self::FCurve25519(pf) => Self::FCurve25519(FCurve25519(pf.0.inverse().unwrap())),
             Self::FBls12381(pf) => Self::FBls12381(pf.invert().unwrap()),
             Self::FBn254(pf) => Self::FBn254(pf.invert().unwrap()),
             Self::IntField(i) => Self::IntField(i.recip()),
@@ -213,8 +287,9 @@ impl FieldV {
     #[inline]
     pub fn is_zero(&self) -> bool {
         match self {
-            Self::FBls12381(pf) => bool::from(pf.is_zero()),
-            Self::FBn254(pf) => bool::from(pf.is_zero()),
+            Self::FCurve25519(pf) => pf.0.is_zero(),
+            Self::FBls12381(pf) => bool::from(ff::Field::is_zero(pf)),
+            Self::FBn254(pf) => bool::from(ff::Field::is_zero(pf)),
             Self::IntField(i) => i.is_zero(),
         }
     }
@@ -224,6 +299,7 @@ impl FieldV {
     pub fn is_one(&self) -> bool {
         use num_traits::One;
         match self {
+            Self::FCurve25519(pf) => pf.0.is_one(),
             Self::FBls12381(pf) => pf.is_one(),
             Self::FBn254(pf) => pf.is_one(),
             Self::IntField(i) => i.i == 1,
@@ -237,6 +313,10 @@ impl FieldV {
     {
         let i = Integer::from(i);
         match ty {
+            FieldT::FCurve25519 => {
+                let i = BigInt::from_bits_be(&i.to_digits(Order::MsfBe));
+                Self::FCurve25519(FCurve25519(Fr::from_bigint(i).unwrap()))
+            }
             FieldT::FBls12381 => Self::FBls12381(FBls12381::from(i)),
             FieldT::FBn254 => Self::FBn254(FBn254::from(i)),
             FieldT::IntField(m) => Self::IntField(IntField::new(i, m)),
@@ -245,6 +325,7 @@ impl FieldV {
 
     fn random(ty: FieldT, mut rng: impl rand::RngCore) -> Self {
         match ty {
+            FieldT::FCurve25519 => Self::FCurve25519(FCurve25519(Standard.sample(&mut rng))),
             FieldT::FBls12381 => Self::FBls12381(FBls12381::random(rng)),
             FieldT::FBn254 => Self::FBn254(FBn254::random(rng)),
             FieldT::IntField(m) => {
@@ -277,6 +358,7 @@ impl FieldV {
 impl Display for FieldV {
     fn fmt(&self, f: &mut Formatter) -> fmt::Result {
         match self {
+            Self::FCurve25519(pf) => pf.0.fmt(f),
             Self::FBls12381(pf) => write!(f, "#f{}m{}", Integer::from(pf), &*F_BLS12381_FMOD),
             Self::FBn254(pf) => write!(f, "#f{}m{}", Integer::from(pf), &*F_BN254_FMOD),
             Self::IntField(i) => i.fmt(f),
@@ -309,6 +391,7 @@ macro_rules! arith_impl {
                         (Self::FBls12381(f1), Self::FBls12381(f2)) => f1.[<$fn _assign>](f2),
                         (Self::FBn254(f1), Self::FBn254(f2)) => f1.[<$fn _assign>](f2),
                         (Self::IntField(i1), Self::IntField(i2)) => i1.[<$fn _assign>](i2),
+                        (Self::FCurve25519(i1), Self::FCurve25519(i2)) => i1.0.[<$fn _assign>](i2.0),
                         (s, o) => panic!("Operation [<$Trait Assign>] on {} and {}", s.ty(), o.ty()),
                     }
                 }
@@ -331,6 +414,7 @@ impl Neg for FieldV {
     type Output = Self;
     fn neg(self) -> Self {
         match self {
+            Self::FCurve25519(pf) => Self::FCurve25519(FCurve25519(pf.0.neg())),
             Self::FBls12381(pf) => Self::FBls12381(pf.neg()),
             Self::FBn254(pf) => Self::FBn254(pf.neg()),
             Self::IntField(i) => Self::IntField(i.neg()),
@@ -363,6 +447,10 @@ impl Into<FieldV> for IntField {
 impl Into<Integer> for FieldV {
     fn into(self) -> Integer {
         match self {
+            FieldV::FCurve25519(f) => {
+                let i: BigInt<4> = f.0.into_bigint();
+                Integer::from_digits(&i.to_bytes_be(), Order::MsfBe)
+            }
             FieldV::FBls12381(f) => Integer::from(&f),
             FieldV::FBn254(f) => Integer::from(&f),
             FieldV::IntField(i) => i.i,
@@ -374,9 +462,22 @@ impl Into<Integer> for FieldV {
 impl Into<Integer> for &FieldV {
     fn into(self) -> Integer {
         match self {
+            FieldV::FCurve25519(f) => {
+                Integer::from_digits(&f.0.into_bigint().to_bytes_be(), Order::MsfBe)
+            }
             FieldV::FBls12381(f) => Integer::from(f),
             FieldV::FBn254(f) => Integer::from(f),
             FieldV::IntField(i) => i.i.clone(),
         }
     }
+}
+
+/// convert BigInt from ark to Integer from rug
+pub fn bigint_to_int<const N: usize>(i: BigInt<N>) -> Integer {
+    Integer::from_digits(&i.to_bytes_be(), Order::MsfBe)
+}
+
+/// convert Integer from rug to BigInt from ark
+pub fn int_to_bigint<const N: usize>(i: Integer) -> BigInt<N> {
+    BigInt::from_bits_be(&i.to_digits(Order::MsfBe))
 }