--- a/src/ir/term/precomp.rs
+++ b/src/ir/term/precomp.rs
@@ -3,9 +3,16 @@
 //! Conceptually, this machinery allows a party with input material for one computation to map it
 //! into input material for another computation.
 
-use fxhash::{FxHashMap, FxHashSet};
+// use std::time::Instant;
 
-use crate::ir::term::*;
+use std::{time::Instant};
+
+use fxhash::{FxHashMap, FxHashSet};
+use itertools::Itertools;
+// use std::time::{Instant};
+
+use crate::{ir::term::*, target::r1cs::R1cs};
+use smallvec::SmallVec;
 
 /// A "precomputation".
 ///
@@ -13,8 +20,9 @@ use crate::ir::term::*;
 #[derive(Debug, Clone, Default, PartialEq, Eq, Serialize, Deserialize)]
 pub struct PreComp {
     /// A map from output names to the terms that compute them.
-    outputs: FxHashMap<String, Term>,
-    sequence: Vec<String>,
+    pub outputs: FxHashMap<String, Term>,
+    /// The sequence of output names
+    pub sequence: Vec<String>,
 }
 
 impl PreComp {
@@ -58,10 +66,46 @@ impl PreComp {
             !drop
         });
     }
-    /// Evaluate the precomputation.
-    ///
-    /// Requires an input environment that binds all inputs for the underlying computation.
-    pub fn eval(&self, env: &FxHashMap<String, Value>) -> FxHashMap<String, Value> {
+
+    /// transform Term to NumTerm
+    pub fn transform_terms(
+        &self,
+        index_cache: &mut TermMap<usize>,
+        term_arr: &mut Vec<NumTerm>,
+    ) {
+
+        for o_name in &self.sequence {
+            let t = self.outputs.get(o_name).unwrap();
+            let mut stack = vec![(false, t.clone())];
+            while let Some((children_pushed, node)) = stack.pop() {
+                if index_cache.contains_key(&node) {
+                    continue;
+                }
+                if children_pushed {
+                    let mut cs_arr = SmallVec::<[usize; 3]>::new();
+                    for child in &node.cs {
+                        cs_arr.push(*index_cache.get(child).unwrap());
+                    }
+                    let new_num_term = NumTerm {
+                        op: NumOp::Op(node.op.clone()),
+                        cs: cs_arr,
+                    };
+                    index_cache.insert(node, term_arr.len());
+                    term_arr.push(new_num_term);
+                } else {
+                    stack.push((true, node.clone()));
+                    for c in &node.cs {
+                        if !index_cache.contains_key(c) {
+                            stack.push((false, c.clone()));
+                        }
+                    }
+                }
+            }
+        }
+    }
+
+    /// function
+    pub fn eval(&self, env: &FxHashMap<String, Value>) -> FxHashMap<String, Value>{
         let mut value_cache: TermMap<Value> = TermMap::new();
         let mut env = env.clone();
         // iterate over all terms, evaluating them using the cache.
@@ -72,6 +116,52 @@ impl PreComp {
         }
         env
     }
+
+    /// eval preprocess
+    pub fn eval_preprocess(
+        &self,
+        r1cs: &R1cs<String>,
+    ) -> (Vec<NumTerm>, FxHashMap<String, usize>) {
+        let mut term_arr = Vec::<NumTerm>::new();
+        let mut index_cache = TermMap::<usize>::new();
+        self.transform_terms(&mut index_cache, &mut term_arr);
+        for term in &mut term_arr {
+            if let NumOp::Op(op) = &term.op {
+                if let Op::Var(o_name, _) = op {
+                    if let Some(o) = self.outputs.get(o_name) {
+                        let term_idx = *index_cache.get(o).unwrap();
+                        term.op = NumOp::Var(term_idx);
+                    }
+                }
+            } else {
+                panic!("Op should not be transformed here!");
+            }
+        }
+        let mut name_idxes = FxHashMap::default();
+        for (_, name) in r1cs.idxs_signals.iter().sorted() {
+            let o = self.outputs.get(name).unwrap();
+            let term_idx = *index_cache.get(o).unwrap();
+            name_idxes.insert(name.clone(), term_idx);
+        }
+        (term_arr, name_idxes)
+    }
+
+    /// real evaluation
+    pub fn real_eval(name_idxes: &FxHashMap<String, usize>, term_arr: &Vec<NumTerm>, env: &FxHashMap<String, Value>) -> FxHashMap<String, Value> {
+        let t1 = Instant::now();
+        let mut val_arr = vec![Option::None; term_arr.len()];
+        for term_idx in 0..term_arr.len() {
+            eval_value_efficient(term_idx, term_arr, &mut val_arr, env);
+        }
+        println!("real eval take {}", t1.elapsed().as_millis());
+        let mut map = FxHashMap::default();
+        for (name, idx) in name_idxes {
+            map.insert(name.clone(), val_arr[*idx].clone().unwrap());
+        }
+        println!("all eval take {}", t1.elapsed().as_millis());
+        map
+    }
+
     /// Compute the inputs for this precomputation
     pub fn inputs_to_terms(&self) -> FxHashMap<String, Term> {
         PostOrderIter::new(term(Op::Tuple, self.outputs.values().cloned().collect()))