… Cargo.lock
… circ_fields/Cargo.lock
… circ_fields/Cargo.toml
… examples/run_zok.rs
… src/ir/term/mod.rs
… src/target/r1cs/bellman.rs
… src/target/r1cs/mod.rs
… third_party/curve25519/Cargo.toml
… third_party/curve25519/LICENSE-APACHE
… third_party/curve25519/LICENSE-MIT
… third_party/curve25519/scripts/fr.ipynb
… third_party/curve25519/scripts/g1.ipynb
… third_party/curve25519/src/constraints/curves.rs
… third_party/curve25519/src/constraints/fields.rs
… third_party/curve25519/src/constraints/mod.rs
… third_party/curve25519/src/curves/mod.rs
… third_party/curve25519/src/curves/tests.rs
… third_party/curve25519/src/fields/fq.rs
… third_party/curve25519/src/fields/fr.rs
… third_party/curve25519/src/fields/mod.rs
… third_party/curve25519/src/fields/tests.rs
… third_party/curve25519/src/lib.rs