## Motivation

The performance of R1CS proof generation is affected by the slow circuit evaluation, this pull request improved the CirC evaluation speed through several changes
- preprocessed the precompute to avoid dfs and hashmap lookup during evaluation
- change the implementation of Array from BTreeMap to im::OrdMap, which makes the cost of updating an array from O(n) to O(logn)
- Extend opt to PreCompute, `ConstantFold`, `Obliv` and `Tuple` is implemented at this point to convert Oblivious array lookup to variable in PreCompute, this reduce the eval cost a lot
- several Serialization traits are implemented 
- ark-ff Curve25519 is integrated to support Spartan, but it seems Spartan integration exists in another PR, so I remove my Spartan integration

Moreover, after preprocessing, the Prover doesn't need to hold `TermData` to evaluate, but the new `NumTerm` instead, which can potentially reduce the cost of serializing the prover data from O(n^2) to O(n), more code is needed to make this work

## Benchmark
run `cargo run --release --example run_zok` to compare the time taken to evaluate a `shaRound` before and after the optimization

The time dropped from 115ms to 23ms, though 13ms of the 23ms is spent on building a hashmap for bellman, which could be avoided

## TODO

- [ ] test cases