@collinzrj 2022-12-20
Hi Alex, sorry for the late reply. I finally found the time to work on this after the finals. I'm going to fix these issues soon.