@alex-ozdemir 2022-12-21
No worries about the delay.

One thing to note: I recently improved serde support in the compiler, and added a ["linear term"](https://github.com/circify/circ/blob/master/src/ir/term/lin.rs) that's similar to `NumTerm`. You might want to look at those changes when you rebase/merge.