@alex-ozdemir 2023-04-11
Hi Colin!

I'm glad you're taking another look at this. Alas, the compiler has changed a fair bit since you started this PR.

Getting it into a mergeable state is likely to require some work. A good starting point is marking sure the PR *could* be merged cleanly into the latest master.

Regardless of whether you do that, thank you for looking at this. Some of the ideas this PR explored (better serialization, a precomputation format that's better for fast evaluation) have already been upstreamed in other ways.

Cheers,
Alex