@alex-ozdemir 2024-06-25
I'm going to close this, as our evaluation machinery looks quite different now.
Thanks for looking at this, Colin.